+++ include/comm_player_manager.h
