+++ platinum.us/main.lsf
@@ -238,7 +238,7 @@ Static main
 	Object main.nef.p/src_unk_02056B30.c.o
 	Object main.nef.p/src_unk_020573FC.c.o
 	Object main.nef.p/src_unk_0205749C.c.o
-	Object main.nef.p/src_unk_02057518.c.o
+	Object main.nef.p/src_comm_player_manager.c.o
 	Object main.nef.p/src_field_comm_manager.c.o
 	Object main.nef.p/src_unk_0205A0D8.c.o
 	Object main.nef.p/src_unk_0205B33C.c.o
