+++ src/comm_player_manager.c
@@ -31,7 +31,7 @@
 #include "unk_0203A378.h"
 #include "unk_0203CC84.h"
 #include "unk_02054D00.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205E7D0.h"
 #include "unk_0205F180.h"
 #include "unk_02061804.h"
