+++ src/field_comm_manager.c
@@ -24,7 +24,7 @@
 #include "unk_0203CC84.h"
 #include "unk_0203E880.h"
 #include "unk_02050A74.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "field_comm_manager.h"
 #include "unk_0205A0D8.h"
 #include "unk_02071D40.h"
