+++ src/overlay005/ov5_021D0D80.c
@@ -60,7 +60,7 @@
 #include "unk_020556C4.h"
 #include "unk_020559DC.h"
 #include "unk_02055C50.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205D8CC.h"
 #include "unk_0205E7D0.h"
 #include "unk_02061804.h"
