+++ src/overlay005/ov5_021D1C30.c
@@ -41,7 +41,7 @@
 #include "unk_02054D00.h"
 #include "unk_020562F8.h"
 #include "unk_02056B30.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "field_comm_manager.h"
 #include "unk_0205A0D8.h"
 #include "unk_0205B33C.h"
