+++ src/overlay023/ov23_0223E140.c
@@ -68,7 +68,7 @@
 #include "unk_020507CC.h"
 #include "unk_020508D4.h"
 #include "unk_02054D00.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0206A8DC.h"
 #include "unk_0206AFE0.h"
 #include "unk_0206CCB0.h"
