+++ src/overlay023/ov23_022416A8.c
@@ -23,7 +23,7 @@
 #include "unk_0203CC84.h"
 #include "unk_020507CC.h"
 #include "unk_02054D00.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0206A8DC.h"
 #include "overlay005/ov5_021F575C.h"
 #include "overlay023/ov23_0223E140.h"
