+++ src/overlay023/ov23_0224340C.c
@@ -49,7 +49,7 @@
 #include "unk_0203CC84.h"
 #include "unk_020507CC.h"
 #include "unk_02054D00.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205E7D0.h"
 #include "unk_02063400.h"
 #include "unk_0206AFE0.h"
