+++ src/overlay023/ov23_02248F1C.c
@@ -35,7 +35,7 @@
 #include "unk_020218BC.h"
 #include "strbuf.h"
 #include "communication_system.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205D8CC.h"
 #include "unk_0205E7D0.h"
 #include "overlay023/ov23_02241F74.h"
