+++ src/overlay023/ov23_0224B05C.c
@@ -47,7 +47,7 @@
 #include "unk_020530C8.h"
 #include "unk_02054D00.h"
 #include "unk_020573FC.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205E7D0.h"
 #include "unk_0205F180.h"
 #include "unk_02063400.h"
