+++ src/overlay023/ov23_0224DC40.c
@@ -37,7 +37,7 @@
 #include "unk_0203CC84.h"
 #include "unk_020507CC.h"
 #include "unk_020573FC.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0206AFE0.h"
 #include "overlay023/ov23_02241F74.h"
 #include "overlay023/ov23_0224DC40.h"
