+++ src/overlay023/ov23_0224F294.c
@@ -49,7 +49,7 @@
 #include "communication_system.h"
 #include "unk_0203CC84.h"
 #include "unk_020530C8.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205E7D0.h"
 #include "unk_0206A780.h"
 #include "unk_0207D3B8.h"
