+++ src/overlay023/ov23_0225128C.c
@@ -40,7 +40,7 @@
 #include "unk_0203CC84.h"
 #include "unk_020507CC.h"
 #include "unk_020508D4.h"
-#include "unk_02057518.h"
+#include "comm_player_manager.h"
 #include "unk_0205E7D0.h"
 #include "unk_0206A780.h"
 #include "unk_0206A8DC.h"
